+++ frontend/src/metabase/forms/hooks/use-form-submit-button/use-form-submit-button.ts
@@ -44,6 +44,7 @@ const useIsRecent = (value: unknown, timeout: number) => {
 };
 
 const getFormStatus = (status: FormStatus, isRecent: boolean): FormStatus => {
+  console.log("getFormStatus status", status);
   switch (status) {
     case "fulfilled":
     case "rejected":
